+++ src/Symfony/Component/Translation/Tests/Writer/TranslationWriterTest.php
@@ -18,22 +18,6 @@ use Symfony\Component\Translation\Writer\TranslationWriter;
 
 class TranslationWriterTest extends TestCase
 {
-    /**
-     * @group legacy
-     * @expectedDeprecation Method Symfony\Component\Translation\Writer\TranslationWriter::writeTranslations() is deprecated since version 3.4 and will be removed in 4.0. Use write() instead.
-     */
-    public function testWriteTranslations()
-    {
-        $dumper = $this->getMockBuilder('Symfony\Component\Translation\Dumper\DumperInterface')->getMock();
-        $dumper
-            ->expects($this->once())
-            ->method('dump');
-
-        $writer = new TranslationWriter();
-        $writer->addDumper('test', $dumper);
-        $writer->writeTranslations(new MessageCatalogue(array()), 'test');
-    }
-
     public function testWrite()
     {
         $dumper = $this->getMockBuilder('Symfony\Component\Translation\Dumper\DumperInterface')->getMock();